submit button now saves the input data and renders a new DogCard added to the end of the DogList component
