+++ lib/data.js
@@ -1,0 +1,22 @@
+export const dogs = [
+  {
+    name: "Anton",
+    birthdate: "14.07.2014",
+    birthplace: "Germany",
+    sex: "male",
+    color: "black",
+    race: "Shepard Labrador Mix",
+    vaccinations: "none",
+    insurances: "none",
+  },
+  {
+    name: "Canella",
+    birthdate: "01.05.2018",
+    birthplace: "Spain",
+    sex: "female",
+    color: "brown",
+    race: "Unknown Mix",
+    vaccinations: "none",
+    insurances: "none",
+  },
+];
