+++ pages/_app.js
@@ -1,25 +1,6 @@
 import GlobalStyle from "../styles";
 import Head from "next/head";
-
-const dogs = [
-  {
-    id: 1,
-    name: "Anton",
-    dateOfBirth: "14.07.2014",
-    placeOfBirth: "Germany",
-    sex: "male",
-    color: "black",
-    breed: "Shepard Labrador Mix",
-  },
-  {
-    name: "Canella",
-    dateOfBirth: "01.05.2018",
-    placeOfBirth: "Spain",
-    sex: "female",
-    color: "brown",
-    breed: "Unknown Mix",
-  },
-];
+import { dogs } from "../lib/data";
 
 export default function App({ Component, pageProps }) {
   return (
